Refactor types into submodules
